+++ packages/profile-metrics-controller/src/index.ts
@@ -18,3 +18,4 @@ export type {
 } from './ProfileMetricsService';
 export { ProfileMetricsService, serviceName } from './ProfileMetricsService';
 export type { ProfileMetricsServiceMethodActions } from './ProfileMetricsService-method-action-types';
+export type { ProfileMetricsControllerMethodActions } from './ProfileMetricsController-method-action-types';