+++ packages/profile-metrics-controller/CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `ProfileMetricsController` contructor now accepts an optional `initialDelayDuration` parameter ([#7624](https://github.com/MetaMask/core/pull/7624))
   - The parameter can be used to override the default time-based delay for the first data collection after opt-in
+- Add `skipInitialDelay()` method to `ProfileMetricsController` ([#7624](https://github.com/MetaMask/core/pull/7624))
+  - The method can be also called through the `ProfileMetricsController:skipInitialDelay` action via messenger
 
 ### Changed